% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.cv.rbridge}
\alias{plot.cv.rbridge}
\title{Plot a 'cv.rbridge' object function}
\usage{
\method{plot}{cv.rbridge}(x, sign.lambda = 1, ...)
}
\arguments{
\item{x}{Design matrix.}

\item{sign.lambda}{Either plot against \code{log(lambda)} (default) or its negative if sign.\code{lambda=-1}.}

\item{...}{Other graphical parameters to plot}
}
\description{
Plots the cross-validation curve, and upper and lower standard deviation curves, as a function of the lambda values used.
}
\author{
Bahadir Yuzbasi, Mohammad Arashi and Fikri Akdeniz \cr Maintainer: Bahadir Yuzbasi \email{b.yzb@hotmail.com}
}
